Code reorganization
