+++ core/src/main/java/platformer/audio/OpenAL.java
@@ -3,6 +3,9 @@ package platformer.audio;
 import org.lwjgl.openal.AL;
 import org.lwjgl.openal.AL10;
 import org.lwjgl.util.WaveData;
+import platformer.audio.types.Ambience;
+import platformer.audio.types.Song;
+import platformer.audio.types.Sound;
 import platformer.debug.logger.Logger;
 import platformer.debug.logger.Message;
 import platformer.utils.ValueEnum;
