+++ core/src/main/java/platformer/model/gameObjects/IntersectionHandler.java
@@ -1,7 +1,7 @@
 package platformer.model.gameObjects;
 
 import platformer.audio.Audio;
-import platformer.audio.Sound;
+import platformer.audio.types.Sound;
 import platformer.model.entities.Direction;
 import platformer.model.entities.enemies.EnemyManager;
 import platformer.model.entities.player.Player;
