+++ core/src/main/java/platformer/model/gameObjects/objects/Shop.java
@@ -1,7 +1,7 @@
 package platformer.model.gameObjects.objects;
 
 import platformer.audio.Audio;
-import platformer.audio.Sound;
+import platformer.audio.types.Sound;
 import platformer.model.entities.player.Player;
 import platformer.model.gameObjects.GameObject;
 import platformer.model.gameObjects.ObjType;
